Migrated TreeGridAxis to ES6 pattern.
